developer-productivity-dashboard/requirements.txt
rich>=13.0.0
GitPython>=3.1.0
PyGithub>=2.0.0
click>=8.0.0
